Add an empty passphrase file
As of `bitnami/redis:6.2.7-debian-11-r18`, the startup script for redis
requires a file containing a passphrase to be provided via the
`REDIS_TLS_KEY_FILE_PASS` environment variable if Redis is running with
TLS enabled. Since we're only using this container in local development
and not in production, we can override by creating an empty file when we
initialize the SSL certificate so that it passes the "Does the file
exist?" check that the redis script is running.