Use Faculty.findOneOrFail to validate entity presence
Instead of having the service method call `Repository.findOneOrFail()`,
it might make more sense to call that in the controller using the
absence info provided in the request body, then pass the resulting
`Absence` entity to the service method.

That way, a missing entity will be correctly captured and handled by the
try/catch block below rather than being swallowed and re-thrown as a
`BadRequestException`(which is not quite correct for a missing entity).

Because the `Absence` entity provided to the service method comes
straight from the database, it is guaranteed to exist - making the call
to `.findOneOrfail` a little redundant in there. However, that can be
removed later...