Add API for getting the room list
The existing API that is getting the room availability was updated so that its path is /api/rooms/availability instead so that this more general retrieval of rooms without availability can have the path /api/rooms. In order to get the room information, we are querying the existing roomListingViewRepository. While we do not technically need the capacity or campus information to build the room dropdown, we can reuse this service for the future room admin tab.