Add RoomSchedule component
This room schedule component includes a combobox that has a filter function that uses the mark-one regex filter function which allows a case insensitive search. It also allows you to search the middle of a word (e.g. if the room is 1414 Massachusetts Avenue 230, you could search "ave" to see this room as one of the search results). Two NoteText components were also added to the menu that show the (for now) hardcoded semester and currently selected room.