Query absenceRepository for test data instead of facultyRepository
The test data was previously being obtained by querying the
facultyRepository, for a faculty member, then destructuring their
absences to get a single absence.

This is somewhat inefficient as it brings unnecessary date into the test
as well as adding overhead with unnecessary `JOIN`s. Instead, it makes
more sense to simply query the absenceRepository directly, which then
allows for a more efficient (and hopefully faster) set of tests