Add RoomMeetingResponse DTO and update existing room availability code to use it
The RoomMeetingResponse DTO has all of the properties that the original RoomResponse DTO had. The existing code related to room availability was updated to use the RoomMeetingResponse DTO, and the code related to the room combobox on the RoomSchedule page uses the RoomResponse DTO, which has the propeties id, name, campus, and capacity.