Ignore optional dependencies
npm's default behavior is to install all optional dependencies by
default, which seems to go against the idea of them being "optional."
Since node-canvas is an optional peer dependency of jsdom, that's also
being pulled in even though it's not supported by arm architectures out
of the box.

As a somewhat blunt solution, we can omit all optional dependencies both
in a local .npmrc (which will be picked up when running npm install in
the project) and in our Dockerfile (which will be picked up when
building the container). That should resolve the issues running the app
on Apple M1 computers without needing to use the translation layer to
run the container in amd64 mode.