Add optional catalogNumber prop to SessionBlock
The purpose of this block is so that the RoomSchedule block headings can display the full catalogNumber as opposed to just the prefix to match the mockup that Vittorio made for the Room Schedule table. If a catalog number is provided, it will display the catalogNumber instead of the prefix.