Add the RoomSchedule component
The RoomSchedule is hard coded to get the data for the academic year 2020. When the tab first loads, no room has been yet selected, so we have a note to indicate that the schedule will show when the user selects a room. When the schedule data exists, the schedule blocks will include the catalog number, title, faculty, day, and time of the course instance.