Remove findOneOrFail call to find absence
Since the controller already retrieves the absence by ID and handles
missing records, this logic can be removed and left to the controller to
deal with.

Otherwise the service method is simply re-implementing what already
exists in the controller. This is especially true where
`NotFoundException`s are being thrown which should _really_ be left
in the controller since they're HTTP exceptions