Add controller integration tests
These tests check the results of calling the room schedule API with invalid terms, years, and room IDs. As opposed to mocking and manually resolving the finding of rooms, we are using the test databases to find a room and using the room id of the room we found to make the API request.