Restructure tests to reduce database queries
Lots of database queries were being run and then the results effectively
discarded and re-fetched in the form of re-issuing the same or a similar
query.

This is somewhat wasteful of CPU cycles, makes the tests slower to run
and (perhaps more importantly) makes the tests harder to read.

Instead, it makes more sense to fetch some data at the beginning that's
more or less what's needed and then edit it as the test suite goes on
to more closely match what's expected.

In this case, that means getting rid of the `noLongerActiveAbsence`
object and instead having `springAbsence` and `fallAbsence`. As well
as more closely mirroring the real application data, this makes things
easier to reason about what absence is being edited