Update OfferedModal to use getInstanceIdentifier utils function
The currentSemester property's calendarYear property was updated to be called academicYear, since the selectedAcademicYear value is truly an academic year provided initially by the metadata context (which is taking the academic year from the config service). The modal uses the currentSemester's year only for display purposes, so it is safe to rename the "calendarYear" to "academicYear" under currentSemester. No changes in requests were made - all changes only affect the display.