Fix Enrollment modal to use getInstanceIdentifier
The currentSemester prop's calendarYear value of the enrollment modal was only being used to populate the header, so it should be safe to rename it properly to academicYear. Like the Offered Modal, it is also pulling the academic year from the metadata context, so it is truly the academic year and not the calendar year.