Fix the display text for InstructorModal
The currentSemester's calendarYear property was only being used to be passed into the instanceIdentifier variable, so I have replaced the original logic by updating the currentSemester to have the academicYear property and use the getInstanceIdentifier utils function. The selectedAcademicYear on the Courses page is the academic year from the metadata context, which is truly the academic year.