Fix InstructorModal test
This commit updates the instructor modal to use the academic year rather than the calendar year. I tried to future proof these tests by calculating the academic year by using the semKey rather than hard coding the academic year so that if someone updates the testTerm to TERM.SPRING in the future, the testAcademicYear will not need to be changed.