Render result access
Making the return value of `renderWithMessaging` accessible in the test
suite(rather than using de-structured functions) makes for tests that
are easier to reason about (i.e: it's easier to see what scope we're
working in whether that's within the whole page, within just a modal
etc.).

It also helps avoid naming collisions when dealing with multiple
`getBy*` functions for multiple different elements.