Add tests to check that the controller works as expected with valid parameters
In order to have a test that checks that a non empty array of data is returned, we must first query for a course instance that has at least one meeting in one room. This will give us a combination of room id, term, and year that we can use to make an api request that should return us with room schedule data (not an empty array).