Add test to check that the faculty are returned in the correct order
This test queries for the course instance id in order to get the existing instructors in the database. Using the existing instructors, we can compare this to the instructors returned as a result from the api in the room schedule and compare that their IDs appear in the expected order (organized by instructor order ascending).