Replace usage of `.where` with `.andWhere`
The usage of `.where` was overriding all other `WHERE` clauses in the
query which means that `UPDATE absence WHERE id IN(:ids)` was instead
being replaced with `UPDATE absence WHERE id <> :absenceId`.

This meant that the code was updating _EVERY_ absence in the db (except
the one whose ID was `:absenceId`) to `ABSENCE_TYPE.PRESENT`, then going
back and updating the one with the ID of `:absenceId` to be whatever
value was provided via `absenceReqInfo.type`

This had the effect of making it _look_ like the code was working in the
browser, but caused the tests to (correctly) fail.

Usage of `.andWhere` correctly tags an `AND WHERE...` clause on the end
and fixes this issue.