Remove `Absence.type` default value
The presence of a default value here masks (what should be) test
failures in such a way that is difficult to debug.

If the `type` column is mistakenly omitted from a query, instead of
having the (correct) value of `undefined`, the field has an alternative
(and valid) value with no obvious source. In fact, it almost appears as
if the database is (for some unknown reason) returning an incorrect
value.