Fix calendarYear value for MeetingModal
The Meeting Modal's room selection component uses the calendar year in order to make a request for the room availability (via getRoomAvailability), we are passing in the calendar year to the MeetingModal component as opposed to the academic year. The term is checked in order to determine whether the calendar year is the same as the academic year (when the term is Spring) or if the calendar year is the year before the academic year (when the term is Fall).