Improve client side validation errors
This reduces the number of different states by storing the errors for each of the form fields in one state object, formErrors, instead of multiple states (one for each field). If any of the fields are invalid, there is one call to setFormErrors in the if block checking that isValid condition is false. This change makes the code a little more concise.