Add the controller and service methods for creating a new room
The service will not allow users to create buildings that exist on other campuses. It will make sure that the campus selected already exists, as new campuses should not be allowed to be created. It also makes sure that the room is not a duplicate of one that already exists. Since rooms are linked to meetings, we do not want users to be able to create duplicate rooms and potentially double-book rooms when assigning meetings to rooms.