Fix service so that we are not passing unnecessary arguments to the .save method
The campus is nested within building, so we should not be passing it separately to the .save method. The campus relations was also set on the building so that we can have access to the campus table from the building. The result was also updated so that it matches what we expect for the RoomAdminResponse.