Update the MetadataContext and CreateRoomModal to allow new building additions
This commit adds the updateCampuses method to the MetadataContext so that we can add new buildings to the metadata as we interact with the create room modal. Before adding to the metadata, we check in the CreateRoomModal that the building does not already exist in the metadata. The updateCampuses method sorts the list of buildings in the campuses alphabetically so that they appear sorted when the user next opens the create room modal and clicks on the buildings dropdown.