Define metadataBuildingIndex outside of the submitForm and check if the campus was filled out before accessing the metadata buildings
The metadataBuildingIndex is being used in a few different spots, and it needs to be defined outside of the submitForm function in order to be accessible in those spots. There is now a conditional that checks whether the campus value was filled in in the form before accessing the metadata. In the case where the user has not selected a campus, we should not be trying to access the buildings of a null campus.