Add tests for the fall absence and spring absence filters
These tests check that the results of the data displayed in the table when changing and not changing the filters. When the filter hasn't been changed, all of the rows in the table should be displayed, whereas the filters should  make it so that only the rows fitting the filter are displayed.