Change sameAs to a self-referential foreign key
The replaces the existing free text field with a foreign key field that
relates back to the course table, creating an "Adjacency List" tree
entity.

We're defining it as a parent-children relationship, where any parent
can have multiple children. However, the actual use case is more like a
sibling-to-sibling relationship. We'll need to tweak our querying to
represent that.