Update get/post/put queries to use new SameAs
When fetching the list of courses, we'll join with the parent course and
use the parent course number in the sameAs field. This is
backwards-compatible with the previous endpoint for populating the
Course Admin.

This modifies the logic so that a request to create or update a course
will expect the uuid for the sameAs parent, which we'll then query from
the database and embed. This is NOT backward-compatible with the
previous client code, as we're expecting a UUID instead of a catalog
number and we'll throw an EntityNotFound error if we can get the course
associated with that UUID.