Checkout version before building
This added step will check out the latest version of our code, as
determined in the pre-publish stage. We need this because the checkout
action will, by default, run on the "main" branch, and we want to be
sure that we're building off of whatever code was just merged by using
the incremented version tag.

Note that we're building both the `_QA` and `_PROD` static bundles off
of the same code, but parsing different environment variable to
set, among other things, the client and server URLs. This is different
from our docker images, where the `:qa` and `:stable` tags track the
develop and main branches, respectively. This is kind of confusing and
may change in the future, but for now I suggest we leave it as is.