Update placeholder text generated by helper function
Originally, the placeholder text was created by using the field value, which tends to be in camel case. Instead, we would like the placeholder text to be in title case. Using the helper function, camelCaseToTitleCase, we can manipulate the field and subfields to have the placeholder text show as expected. The reason why we are bringing in the subField is for properties nested under the fall and spring semesters. For now, we only have one text filter under the fall and spring semesters (instructor column), but we may want to add more text filters in the future. By incorporating the subField into the placeholder text, there is a differentiation between different properties nested under fall and spring.