Use correct fetch-depth parameter
The `fetch-depth: 0` option is required to check out the entire git
history for the project, so that we can then checkout the specific
version we want to build. However, I used `fetch_depth` (with an
underscore), which is not a supported property. The action then default
to `fetch-depth: 1`, which doesn't include the tag we want to checkout,
so the run failed.

In the future we may want to change it to just checkout the specific
version from the beginning, but in the interest of creating the smallest
change possible we can just fix this line for now.