Replace typeorm subscriber with controller code
In order to expidite the patching of the bug, the subscriber has been replaced
with code in the Nest controller. This is due to the fact that changes to how
records are created would have knock-on impacts on the population service and
ETL script which would potentially take time to patch and review.

Therefore, #625 has been opened to track this tech debt and will be revisited
at a later time