Compare test data with `''` instead of null
The `physicsCourseResponse` object is intended to represent the data
returned from the server. This means that for courses where the
`sameAs` field has no value, `null` is returned instead.

In this case, however it is being compared against an HTML text field
which can only return string values. In the code, we use some logic to
convert `''` to `null`, however that runs after the form submission.

This means that a course with no `sameAs` value will have `''` as the
value of it's`sameAs` field, therefore it makes sense to simply compare
that with empty string instead of `null` (which is what
physicsCourseResponse.sameAs` is set to).