Define the FacultyFilterState and SemesterFIlterState as types
Instead of interfaces, the FacultyFilterState and SemesterFilterState make more sense as types, because they are describing the shape of data for the filter fields rather than describing abstract behavior from implementing a class. These definitions were also moved back to the FacultyPage, since they will really only be used for the FacultyPage.