Update the filter labels to use regex
The test labels are now using regex to identify the labels using the word "filter" and the identifying filtering characteristics (e.g. "area" for the area label). Hopefully, this is future proofing the tests in the case that the the filter labels on the Faculty table change in the future.