Remove spread operator
Usage of the spread operator here was causing the new absence to be
(incorrectly) linked back to the previous absence record which then
resulted in the absence record being updated (instead of a new one
created). This then cascaded into test failures as other values in the
database were different than expected as the code modified existing data
rather than creating new records.