Add a one-time fix for missing academic year
When we copied the data from the old Course Planner there were no
semesters for the 2025-2026 academic year. This would be the fourth year
of our four-year plan in 2023, which causes two problems:

1. The admins are unable to add anything for this year, leaving the
   columns blank in the interface.
2. There will no "previous year" data to copy over when we try to add
   the next new year (2026-2027) in June.

To resolve this, we can add a hard-coded call to addAcademicYear that
will add the missing 2026 year. Because this won't run if the year
already exists, this shouldn't cause problems if it attempts to call the
function twice. However, we should probably just revert this commit and
drop the extra code after the has been populated.