Add new academic years any time after June
I think originally we were concerned about running the process too many
times and potentially slowing down the container startup, but I think
that's a less significant concern that potentially skipping a year if
the container doesn't restart in June. I'm more confident that we'll
restart the container at least once during the back half of the year,
and we can monitor logs from that period to make sure we're not causing
unintentional slowdowns or other issues.