WIP: Update webpack configs
While it is compatible with Node 18, Webpack 5 also changes a number of
configuration options and has other breaking changes in plugins that
need to be addressed.

One of particular note is the html-webpack-root-plugin, which is no
longer maintained and is incompatible with the newest
html-webpack-plugin. This plugin was only adding a <div id="root"> tag
to the body of the generated index.html, so the easiest workaround is for
us to provide a template index.html that includes that div.

This is still a WIP, as the webpack-start-server-plugin also appears to
be broken; A resolution for that still needs to be found.