Streamline logic in ReportDownloadModal
Instead of parsing strings, we are now using a headerText prop to pass in the text of the modal header. The new getReport prop will be passed in as either the getFacultyReport or getCourseReport. This improvement makes it so that if we ever have different types of report in the future, we can just add in a new API method and header text instead of parsing strings and adding in conditionals.