WIP: Update nestjs, typeorm packages
This updates a few additional packages in order to resolve cascading peer
dependency issues and bring us up to the latest versions.

Both sets of packages have breaking changes that need to be addressed.

For NestJS, there are migrations guides for going from v7 to v8, then v8
to v9:

https://docs.nestjs.com/v8/migration-guide
https://docs.nestjs.com/migration-guide

And the release notes for TypeORM 0.3.0 include several deprecations
and breaking changes that will need to be addressed:

https://github.com/typeorm/typeorm/releases/tag/0.3.0