Add state in the ScheduleView to prevent course children popups
Now that the logic related to comparing instance IDs with the course sameAs id has been added, this was causing the child courses' popovers to show when the parent was clicked. This extra logic is checking the course number and course prefix in order to prevent the popovers of the child courses to show when the parent gets clicked.