Show "Check Availability" for all non-seas rooms
Any rooms not managed by SEAS were assumed to be FAS. However, given the
possible usage of HBS, Longwood etc. rooms in the future....adding a special
case for each campus would quickly become unmanagable. Add to that the fact
that HBS is _technically_ in Allston.

Therefore, any non-seas rooms are now part of a ficticious "Non-SEAS" campus
in the campus table rather than creating special cases for each Harvard school.
This still allows the room to be booked, but under the proviso of "We don't
have any availability for this room, so you should double check with room book"