Fix the fading behavior in the SchedulePage
To narrow the scope of what is considered truthy for popoverInBlock and isSelected, we are using the clickedCourseInfo state to compare the prefix and number as well. The instanceId is no longer enough to use as a comparator because of the sameAs behavior that we added. For example, if a course has a parent, we used the parent instances instead of that course's instances.