Disable "sameAs" input for parent courses
In order to prevent the creation of a sprawling hirearchy of
parent/child courses, it's important to make sure that any course that
another course is pointing to as it's parent cannot have it's "sameAs"
field set.

That is to say, a course that is already the parent of another course,
cannot also be a child of another course. To prevent this, the select is
disabled in this instance.