Fix the client test to account for sameAs field
Now that the associated (via the sameAs field) courses show in the Faculty table, we must account for them displaying. The test was failing, because the AM test data had a non-empty sameAs value, so more data was showing under the Courses column than previously anticipated. The helper function formatSameAs was added to update the format of the sameAs value to how the cell data is actually formatted.