Allow sameAs value to be unset
According to this Stack Overflow issue (https://github.com/typeorm/typeorm/issues/2934), TypeORM considers an undefined value as being a no-op, meaning that it will not update the value for the property in the database. To solve this issue of the sameAs value not being able to be unset, we should explicitly set the sameAs value to null so that users can actually remove the sameAs value when they are editing the course.