Add a cleanup call in beforeEach
I spent some time trying to isolate and debug some of the test failures
that I was seeing locally. In general, there seemed to be a lot errors
that were starting to appear while testing the "discard unsaved changes"
features in some of the modals. There was some unexpected behavior
around these errors, like TypeORM errors cropping up while
testing client-side code.

That usually indicates an asynchronous test leak, as a rejected promise
will often show up in a different test the one that initiated the
promise. I couldn't find any obvious cases of that, but I did stumble
into another potential fix: cleaning up the react rendering
before running the test.

This shouldn't actually be doing anything, as react-testing-library will
automatically inject a call to cleanup in an afterEach block. And in
practice, I could achieve the same result just by using the screen.debug
method in a beforeEach -- In reality, I think adding the beforeEach
might just be introducing some slight amount of delay that enables the
tests to work as expected.

Whatever the case, this cleanup call does seem to be fixing the
tests locally for me, so I'll push to CI and see how it goes...