Add Population Module
I suspect that some of the errors in these tests are appearing because
the PopulationModule hasn't been added to the TypeORM module. Without
that module imported, the test database is not being cleared out between
test runs, so some data is persisting in state and causing problems.

I should not that, additionally, there are some tests that don't seem to
have well-defined assertions here, e.g.

  context('when required fields are provided', function () {
    it('does not display an error', async function () {
      ...
      ...
      ...
      const submitButton = getByText('Submit');
      fireEvent.click(submitButton);
      await wait(() => !queryByText('required fields', { exact: false }));
    });
  });

Ideally that could be redone with a "waitForElementToBeRemoved", or
"queryBy" for a more specific string that could be compared to null.
That's mainly a problem because the tests could be reporting as
successful when really they just haven't run any assertions.