Improve offered tests
The original tests were not explicitly defining the original offered value, potentially causing typeorm to find retired courses and trying to unretire them. This could be what is potentially causing the tests to fail in the build. These changes explicitly define the original and updated retired values so that the original course instance selected is not one that unexpectedly has a retired value for the offered field.