Make the offered e2e tests stricter by ensuring the original course instance is not retired
This commit adds some logic to make sure that the course that we choose to update is not linked to a course instance that is already retired for the specific academic year and term. This is following a hunch that the tests may be failing periodically due to the test trying to updated a retired course instance. Also, instead of hardcoding a course catalog number that may not exist in the other test database, we are using typeorm to find a course that definitely exists. This commit also splits a test checking the future offered value of course instances into two tests, because I noticed that this test was bundling a large amount of checking into one test. It was checking the offered value for the spring semester, navigating to the new academic year, and then checking the values of the fall and spring semesters of the new academic year.