Fix Schedule end to end test
This test was originally sporadically failing. The test was improved by looking for a test course instance with a non-null sameAs value in order to find a parent-child course pair as opposed instead of taking a course and updating the sameAs value in the course admin modal. This takes out several unnecessary steps and cuts down on the amount of time the test needs to run. One important note is to make sure that the test course instance has an isSEAS value of IS_SEAS.Y. Otherwise, the course will not show in the schedule - this was a blocker that took a while to figure out.